+++ shared/utils/index.ts
@@ -5,9 +5,4 @@ export * from "./makeClickableUrl";
 export * from "./preferences";
 export * from "./rides";
 export * from "./transformRideData";
-
-export const flattenQuery = (param?: string[] | string | undefined): string => {
-  const makeString = (s?: string): string => s || "";
-
-  return Array.isArray(param) ? makeString(param[0]) : makeString(param);
-};
+export * from "./general";